Move show-tree tests and add new tests
